Moved process_data back into dish_image and fix_image
